added makefile, added errs package, handling errors in task service and user service, handling errors in resolvers, fixes in repositories

+++ internal/domain/task/service.go
@@ -1,5 +1,7 @@
 package task
 
+import "github.com/msoft-g1/todo-list-backend/internal/errs"
+
 // Service defines task service.
 type Service struct {
 	repo Repository
@@ -12,16 +14,24 @@ func NewService(repo Repository) *Service {
 
 // CreateTask creates a new task.
 func (s *Service) CreateTask(input *CreateInput) (*Task, error) {
-	defCompleted := false
+	completed := false
 	return s.repo.Create(&Task{
 		Description: input.Description,
-		Completed:   &defCompleted,
+		Completed:   &completed,
 		UserID:      input.UserID,
 	})
 }
 
-// UpdateTask updates the task with the specified id.
-func (s *Service) UpdateTask(id uint, input *UpdateInput) (*Task, error) {
+// UpdateTask updates the task with the specified id. The userID is the user who is requesting
+// the update, if that userID does not match with userID in the task, an error is returned.
+func (s *Service) UpdateTask(id uint, userID uint, input *UpdateInput) (*Task, error) {
+	t, err := s.GetTaskByID(id)
+	if err != nil {
+		return nil, err
+	}
+	if t.UserID != userID {
+		return nil, errs.New(errs.CodeAccessDenied, "Solo el usuario dueño de esta tarea puede editarla")
+	}
 	return s.repo.Update(&Task{
 		ID:          id,
 		Description: input.Description,
@@ -29,8 +39,16 @@ func (s *Service) UpdateTask(id uint, input *UpdateInput) (*Task, error) {
 	})
 }
 
-// DeleteTask deletes the task with the specified id.
-func (s *Service) DeleteTask(id uint) error {
+// DeleteTask deletes the task with the specified id. The userID is the user who is requesting
+// the delete, if that userID does not match with userID in the task, an error is returned.
+func (s *Service) DeleteTask(id uint, userID uint) error {
+	t, err := s.GetTaskByID(id)
+	if err != nil {
+		return err
+	}
+	if t.UserID != userID {
+		return errs.New(errs.CodeAccessDenied, "Solo el usuario dueño de esta tarea puede borrarla")
+	}
 	return s.repo.Delete(id)
 }
